jumbo_scraper/pyproject.toml
[tool.poetry]
name = "product-scraper-api"
version = "0.1.0"
description = "A high-performance API for scraping product information from websites with Redis caching"
authors = ["Your Name <your.email@example.com>"]

[tool.poetry.dependencies]
python = "^3.9"
fastapi = "^0.95.2"
uvicorn = "^0.22.0"
pydantic = "^1.10.8"
playwright = "^1.34.0"
beautifulsoup4 = "^4.12.2"
prometheus-client = "^0.16.0"
python-dotenv = "^1.0.0"
aioredis = "^2.0.1"
redis = "^4.5.5"
starlette = "^0.27.0"
aiohttp = "^3.8.4"

[tool.poetry.dev-dependencies]
pytest = "^7.3.1"
pytest-asyncio = "^0.21.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"